PDF Parser/Dense Retriever/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
appdirs==1.4.4
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
backoff==2.2.1
beartype==0.15.0
beautifulsoup4==4.12.3
bleach==6.2.0
bokeh==3.6.2
boto3==1.35.76
botocore==1.35.76
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cohere==5.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
deepsearch-glm==0.26.2
Deprecated==1.2.15
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docling==2.8.3
docling-core==2.7.1
docling-ibm-models==2.0.7
docling-parse==2.1.2
docstring_parser==0.16
docutils==0.21.2
easyocr==1.7.2
effdet==0.4.1
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.0
executing==2.1.0
fastavro==1.9.7
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.55.2
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.10.0
ftfy==6.3.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-pubsub==2.27.1
google-cloud-vision==3.8.1
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h3==4.1.2
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
instructor==1.2.6
iopath==0.1.10
ipython==8.30.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_bokeh==4.0.5
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-openai==0.2.11
langchain-text-splitters==0.3.2
langdetect==1.0.9
langsmith==0.1.147
layoutparser==0.3.4
lazy_loader==0.4
linkify-it-py==2.0.3
litellm==1.44.28
llama-index-core==0.10.68.post1
llama-index-readers-pathway==0.1.3
llama-index-retrievers-pathway==0.1.3
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==1.57.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.12
packaging==23.2
pandas==2.2.3
panel==1.5.4
param==2.1.1
parso==0.8.4
pathway==0.16.0
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pi_heif==0.21.0
pikepdf==9.4.2
pillow==10.4.0
portalocker==3.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
pypandoc==1.14
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.19
python-oxmsg==0.0.1
python-pptx==1.0.2
python-sat==1.8.dev13
pytz==2024.2
pyviz_comms==3.0.3
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
replicate==1.0.4
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
s3transfer==0.10.4
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlglot==10.6.1
stack-data==0.6.3
surya-ocr==0.4.14
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
timm==1.0.12
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
triton==3.1.0
typer==0.12.5
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
unstructured==0.16.9
unstructured-client==0.28.1
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.13
wrapt==1.17.0
xlrd==2.0.1
XlsxWriter==3.2.0
xyzservices==2024.9.0
yarl==1.18.3
zipp==3.21.0